Moving to new input data format. Cleaning up __init__ (#11)
* Moving to new input data format. Cleaning up __init__
* Trying to fix a py27 failure caused by my refactor.
