+++ policyuniverse/action_categories.py
@@ -1,39 +1,13 @@
 import json
 import os
 from collections import defaultdict
-from policyuniverse import action_categories
+from policyuniverse import _action_categories
 
-
-def reformat_console_data(filename):
-    """
-    The AWS Console has a ~3.5MB file mapping actions to 
-    groups (and other info).
-
-    URL is:
-    https://console.aws.amazon.com/iam/api/policies/arn:aws:iam::aws:policy%C2%B6AdministratorAccess/version/v1/policysummary
-
-    Must be authenticated, and probably also requires CSRF token, so I typically
-    just pull the contents from the Network tab of the Chrome inspector.
-    
-    This method will reformat that into a structure containing only
-    the data we care about.
-    """
-    administrator_access = json.load(open(filename, 'r'))
-    groups = defaultdict(list)
-    for service in administrator_access['serviceDetails']:
-        prefix = service['servicePrefix']
-        for action in service['allowedActions']:
-            action_name = '{prefix}:{name}'.format(prefix=prefix, name=action['name'])
-            category = translate_aws_action_groups(action['actionGroups'])
-            groups[action_name.lower()] = category
-
-    with open(filename, 'w') as outfile:
-        json.dump(groups, outfile, indent=2, sort_keys=True)
 
 def translate_aws_action_groups(groups):
     """
     Problem - AWS provides the following four groups:
-        - Permissiosn
+        - Permissions
         - ReadWrite
         - ListOnly
         - ReadOnly
@@ -63,6 +37,19 @@ def translate_aws_action_groups(groups):
         return 'DataPlaneMutating'
     return 'Unknown'
 
+
+def build_action_categories_from_service_data(service_data):
+    action_categories = dict()
+    for service_name in service_data:
+        service_body = service_data[service_name]
+        prefix = service_body['prefix']
+        service_actions = service_body['actions']
+        for service_action, service_action_body in service_actions.items():
+            key = '{}:{}'.format(prefix, service_action.lower())
+            action_categories[key]=service_action_body['calculated_action_group']
+    return action_categories
+
+
 def categories_for_actions(actions):
     """
     Given an iterable of actions, return a mapping of action groups.
@@ -78,7 +65,7 @@ def categories_for_actions(actions):
     groups = defaultdict(set)
     for action in actions:
         service = action.split(':')[0]
-        groups[service].add(action_categories.get(action))
+        groups[service].add(_action_categories.get(action))
     return groups
 
 def actions_for_category(category):
@@ -92,7 +79,7 @@ def actions_for_category(category):
         set of matching actions
     """
     actions = set()
-    for action, action_category in action_categories.items():
+    for action, action_category in _action_categories.items():
         if action_category == category:
            actions.add(action)
     return actions
