+++ policyuniverse/tests/test_expander_minimizer.py
@@ -21,15 +21,15 @@
 """
 import unittest
 import copy
-from policyuniverse import expand_policy
-from policyuniverse import minimize_policy
-from policyuniverse import expand_minimize_over_policies
-from policyuniverse import get_actions_from_statement
-from policyuniverse import all_permissions
-from policyuniverse import minimize_statement_actions
-from policyuniverse import _get_prefixes_for_action
-from policyuniverse import _expand_wildcard_action
-from policyuniverse import _get_desired_actions_from_statement
+from policyuniverse.expander_minimizer import expand_policy
+from policyuniverse.expander_minimizer import minimize_policy
+from policyuniverse.expander_minimizer import expand_minimize_over_policies
+from policyuniverse.expander_minimizer import get_actions_from_statement
+from policyuniverse.expander_minimizer import all_permissions
+from policyuniverse.expander_minimizer import minimize_statement_actions
+from policyuniverse.expander_minimizer import _get_prefixes_for_action
+from policyuniverse.expander_minimizer import _expand_wildcard_action
+from policyuniverse.expander_minimizer import _get_desired_actions_from_statement
 
 
 WILDCARD_ACTION_1 = "swf:res*"
