+++ policyuniverse/tests/test_statement.py
@@ -294,7 +294,7 @@ statement26 = dict(
 class StatementTestCase(unittest.TestCase):
     def test_statement_effect(self):
         statement = Statement(statement01)
-        self.assertEquals(statement.effect, 'Allow')
+        self.assertEqual(statement.effect, 'Allow')
 
     def test_statement_not_principal(self):
         statement = Statement(statement01)
@@ -312,54 +312,54 @@ class StatementTestCase(unittest.TestCase):
 
     def test_statement_principals(self):
         statement = Statement(statement02)
-        self.assertEquals(statement.principals, set(['arn:aws:iam::012345678910:root']))
+        self.assertEqual(statement.principals, set(['arn:aws:iam::012345678910:root']))
 
         statement = Statement(statement03)
-        self.assertEquals(statement.principals, set(['arn:aws:iam::012345678910:root']))
+        self.assertEqual(statement.principals, set(['arn:aws:iam::012345678910:root']))
 
         statement = Statement(statement04)
-        self.assertEquals(statement.principals, set(['arn:aws:iam::012345678910:root']))
+        self.assertEqual(statement.principals, set(['arn:aws:iam::012345678910:root']))
 
         statement = Statement(statement05)
-        self.assertEquals(statement.principals, set(['arn:aws:iam::012345678910:root', 'lambda.amazonaws.com']))
+        self.assertEqual(statement.principals, set(['arn:aws:iam::012345678910:root', 'lambda.amazonaws.com']))
 
         statement = Statement(statement06)
-        self.assertEquals(statement.principals, set(['lambda.amazonaws.com']))
+        self.assertEqual(statement.principals, set(['lambda.amazonaws.com']))
 
         statement_wo_principal = dict(statement06)
         del statement_wo_principal['Principal']
         statement = Statement(statement_wo_principal)
-        self.assertEquals(statement.principals, set([]))
+        self.assertEqual(statement.principals, set([]))
 
     def test_statement_conditions(self):
         statement = Statement(statement07)
-        self.assertEquals(statement.condition_arns, set(['arn:aws:iam::012345678910:role/SomeTestRoleForTesting']))
+        self.assertEqual(statement.condition_arns, set(['arn:aws:iam::012345678910:role/SomeTestRoleForTesting']))
 
         statement = Statement(statement08)
-        self.assertEquals(statement.condition_arns,
+        self.assertEqual(statement.condition_arns,
             set(['arn:aws:iam::012345678910:role/SomeTestRoleForTesting',
             'arn:aws:iam::012345678910:role/OtherRole']))
 
         statement = Statement(statement10)
-        self.assertEquals(statement.condition_accounts, set(['012345678910', '123456789123']))
+        self.assertEqual(statement.condition_accounts, set(['012345678910', '123456789123']))
 
         statement = Statement(statement11)
-        self.assertEquals(statement.condition_accounts, set(['012345678910', '123456789123']))
+        self.assertEqual(statement.condition_accounts, set(['012345678910', '123456789123']))
 
         statement = Statement(statement12)
-        self.assertEquals(statement.condition_arns, set(['arn:aws:iam::012345678910:role/Admin']))
-        self.assertEquals(statement.condition_accounts, set(['012345678910']))
-        self.assertEquals(statement.condition_userids, set(['AROAI1111111111111111:*']))
-        self.assertEquals(statement.condition_cidrs, set(['123.45.67.89', '10.0.7.0/24', '172.16.0.0/16']))
-        self.assertEquals(statement.condition_vpcs, set(['vpc-111111']))
-        self.assertEquals(statement.condition_vpces, set(['vpce-111111']))
+        self.assertEqual(statement.condition_arns, set(['arn:aws:iam::012345678910:role/Admin']))
+        self.assertEqual(statement.condition_accounts, set(['012345678910']))
+        self.assertEqual(statement.condition_userids, set(['AROAI1111111111111111:*']))
+        self.assertEqual(statement.condition_cidrs, set(['123.45.67.89', '10.0.7.0/24', '172.16.0.0/16']))
+        self.assertEqual(statement.condition_vpcs, set(['vpc-111111']))
+        self.assertEqual(statement.condition_vpces, set(['vpce-111111']))
 
         statement = Statement(statement13)
-        self.assertEquals(statement.condition_arns, set(['arn:aws:iam::012345678910:role/Admin']))
-        self.assertEquals(len(statement.condition_userids), 0)
+        self.assertEqual(statement.condition_arns, set(['arn:aws:iam::012345678910:role/Admin']))
+        self.assertEqual(len(statement.condition_userids), 0)
 
         statement = Statement(statement23)
-        self.assertEquals(statement.condition_accounts, set(['222222222222']))
+        self.assertEqual(statement.condition_accounts, set(['222222222222']))
 
     def test_statement_internet_accessible(self):
         self.assertTrue(Statement(statement14).is_internet_accessible())
